Binary Search
This search algorithm takes a collection of elements that is already sorted by ignoring half of the elements after just one comparison. Then compare x with the middle element.
If x matches with the middle element, we return the mid index. Else if x is greater than the mid element, then x lie in the right (greater) half subarray after the mid element. Then we apply the algorithm again for the right half. Else if x is smaller, the target x lie in the left (lower) half. So we apply the algorithm for the left half.